requirements.txt
streamlit
flask
requests
python-dotenv
sendgrid
google-auth
google-auth-oauthlib
google-api-python-client
flask-cors


